added unit handling

+++ BookStoreAlonso/src/control/ActionListenerDelete.java
@@ -1,9 +1,16 @@
 package control;
 
+import java.awt.Image;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+import javax.swing.Icon;
+import javax.swing.ImageIcon;
 import javax.swing.JOptionPane;
+import javax.swing.JSpinner;
+import javax.swing.SpinnerNumberModel;
+
+import tools.DialogBookStore;
 
 public class ActionListenerDelete implements ActionListener {
 	private BookStoreController bookStore;
@@ -17,12 +24,31 @@ public class ActionListenerDelete implements ActionListener {
 
 	@Override
 	public void actionPerformed(ActionEvent e) {
-		bookStore.deleteBook(viewControl.getIsbnSelected(bookStore));
+		int selection = DialogBookStore.optionToDelete();
+		switch (selection) {
+		case 0:
+			selection = DialogBookStore.amountToDelete();
+			selection = bookStore.checkUnits(viewControl.getIsbnSelected(bookStore),selection);
+			if (selection > 0)
+				bookStore.eraseDrives(viewControl.getIsbnSelected(bookStore), selection);
+
+			break;
+		case 1:
+			selection = DialogBookStore.deleteWarning(viewControl.getIsbnSelected(bookStore));
+
+			if (selection == JOptionPane.OK_OPTION) {
+				bookStore.deleteBook(viewControl.getIsbnSelected(bookStore));
+				if (bookStore.getSize() < 1)
+					viewControl.controlStateButtons(viewControl.getBtnSearch());
+				JOptionPane.showMessageDialog(null, "Book Deleted");
+
+			}
+			break;
+
+		default:
+		}
+
 		viewControl.fillTable(bookStore);
-		if (bookStore.getSize() < 1)
-			viewControl.controlStateButtons(viewControl.getBtnSearch());
-		;
-		JOptionPane.showMessageDialog(null, "Book Deleted");
 
 	}
